Prepping the site backend
Wrote some of the code for the site to handle saving & loading of the data for the sites, which will be in its own folder. (NOTE: need to fix the names of the file loaded, since it only goes into the directory. Will need to access the "general.json" file in the dir.