Finished the loading of the json files.
For the most part, loading the json file appears to work perfectly. The unloading from the memory has been untested tho.
The Json files are not auto created though, which will be done in the "Admin" area in its own view. When a site is created, it should make all the json files. On the server's startup, it will check to make sure all files exist for everyone.

Later, I'll work on the way to actually add sites to the website, then allowing site owners to change settings, then the way the pages display themselves.